add a more complex example

+++ Assets/Scripts/Curriculum/CurriculumChallenge.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class CurriculumChallenge : MonoBehaviour
 {
     public int Level { get; set; }
-    public int Target { get; set; }
+    public double Target { get; set; }
 
     public string[] RequiredOperations { get; set; }
     public string[] ExtraOperations { get; set; }
